# server/admin/docs/swagger.yaml
basePath: /api/v1/admin
definitions:
  admin.ErrorResponse:
    properties:
      error:
        example: resource not found
        type: string
    type: object
...
  store.AgentDefinition:
    properties:
      description:
        type: string
      id:
        type: string
      llm:
        $ref: '#/definitions/store.BackendRef'
      mcpServers:
        items:
          type: string
        type: array
      memory:
        $ref: '#/definitions/store.MemoryRef'
      name:
        type: string
      systemPrompt:
        type: string
      systemPromptSuffix:
        type: string
      telegram:
        $ref: '#/definitions/store.TelegramConfig'
      transcription:
        $ref: '#/definitions/store.BackendRef'
      tts:
        $ref: '#/definitions/store.TTSRef'
    type: object
  store.BackendDefinition:
    properties:
      apiKey:
        type: string
      name:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  store.BackendRef:
    properties:
      backend:
        type: string
      model:
        type: string
    type: object
...
  store.MCPServer:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        type: string
      endpoint:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      headers:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      systemPrompt:
        type: string
      type:
        type: string
      workDir:
        type: string
    type: object
...
  store.MemoryRef:
    properties:
      longTerm:
        type: string
      session:
        type: string
    type: object
  store.TTSRef:
    properties:
      backend:
        type: string
      model:
        type: string
      speed:
        type: number
      voice:
        type: string
    type: object
  store.TelegramConfig:
    properties:
      allowedChats:
        items:
          type: integer
        type: array
      allowedUsers:
        items:
          type: integer
        type: array
      enabled:
        type: boolean
      responseMode:
        type: string
      token:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Administration API for managing Magec multi-agent voice assistant resources.
  title: Magec Admin API
  version: "1.0"
paths:
  /agents:
    get:
      description: Returns all configured agents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.AgentDefinition'
            type: array
      summary: List agents
      tags:
      - agents
    post:
      consumes:
      - application/json
      description: Creates a new agent definition
      parameters:
      - description: Agent definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.AgentDefinition'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.AgentDefinition'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Create agent
      tags:
      - agents
  /agents/{id}:
    delete:
      description: Deletes an agent by ID
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Delete agent
      tags:
      - agents
    get:
      description: Returns an agent by its unique ID
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.AgentDefinition'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Get agent
      tags:
      - agents
    put:
      consumes:
      - application/json
      description: Updates an agent by ID
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Agent definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.AgentDefinition'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.AgentDefinition'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Update agent
      tags:
      - agents
  /agents/{id}/mcps:
    get:
      description: Returns the MCP servers linked to a specific agent
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.MCPServer'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: List agent MCPs
      tags:
      - agents
  /agents/{id}/mcps/{name}:
    delete:
      description: Removes the association between an MCP server and an agent
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: MCP server name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Unlink MCP from agent
      tags:
      - agents
    put:
      description: Associates a global MCP server with an agent
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: MCP server name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Link MCP to agent
      tags:
      - agents
  /backends:
    get:
      description: Returns all configured AI backends
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.BackendDefinition'
            type: array
      summary: List backends
      tags:
      - backends
    post:
      consumes:
      - application/json
      description: Creates a new AI backend
      parameters:
      - description: Backend definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.BackendDefinition'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.BackendDefinition'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Create backend
      tags:
      - backends
  /backends/{name}:
    delete:
      description: Deletes a backend by name
      parameters:
      - description: Backend name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Delete backend
      tags:
      - backends
    get:
      description: Returns a backend by its unique name
      parameters:
      - description: Backend name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.BackendDefinition'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Get backend
      tags:
      - backends
    put:
      consumes:
      - application/json
      description: Updates a backend by name
      parameters:
      - description: Backend name
        in: path
        name: name
        required: true
        type: string
      - description: Backend definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.BackendDefinition'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.BackendDefinition'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Update backend
      tags:
      - backends
...
schemes:
- http
swagger: "2.0"